Renamed service folder for consistency
