+++ test/services/MandationStatusServiceSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package service
+package services
 
 
 import org.scalatest.EitherValues
