+++ test/services/mocks/MockUpdateContactPreferenceService.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package service.mocks
+package services.mocks
 
 import org.mockito.ArgumentMatchers
 import org.mockito.Mockito._
@@ -22,20 +22,30 @@ import org.scalatest.Suite
 import org.scalatestplus.mockito.MockitoSugar
 import uk.gov.hmrc.http.HeaderCarrier
 import httpparsers.UpdateVatSubscriptionHttpParser.UpdateVatSubscriptionResponse
-import models.post.EmailPost
-import models.User
-import services._
+import models.{User, VatCustomerInformation}
+import models.post.CommsPreferencePost
+import services.UpdateContactPreferenceService
 
 import scala.concurrent.{ExecutionContext, Future}
 
-trait MockUpdateEmailService extends MockitoSugar {
+trait MockUpdateContactPreferenceService extends MockitoSugar {
   self: Suite =>
 
-  val mockUpdateEmailService: UpdateEmailService = mock[UpdateEmailService]
+  val mockUpdateContactPreferenceService: UpdateContactPreferenceService = mock[UpdateContactPreferenceService]
 
-  def mockUpdateEmail(newPPOBEmail: EmailPost)(response: Future[UpdateVatSubscriptionResponse]): Unit = {
-    when(mockUpdateEmailService
-      .updateEmail(ArgumentMatchers.eq(newPPOBEmail),ArgumentMatchers.any[Boolean])(
+  def mockUpdateContactPreference(newContactPref: CommsPreferencePost)(response: Future[UpdateVatSubscriptionResponse]): Unit = {
+    when(mockUpdateContactPreferenceService
+    .updateContactPreference(ArgumentMatchers.eq(newContactPref), ArgumentMatchers.any[Boolean])(
+      ArgumentMatchers.any[User[_]],
+      ArgumentMatchers.any[HeaderCarrier],
+      ArgumentMatchers.any[ExecutionContext]
+    )
+    ).thenReturn(response)
+  }
+
+  def mockUpdatePreferenceAndEmail()(response: Future[UpdateVatSubscriptionResponse]): Unit = {
+    when(mockUpdateContactPreferenceService
+      .updatePreferenceAndEmail(ArgumentMatchers.any[String], ArgumentMatchers.any[VatCustomerInformation])(
         ArgumentMatchers.any[User[_]],
         ArgumentMatchers.any[HeaderCarrier],
         ArgumentMatchers.any[ExecutionContext]
